engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parityscope"
version = "0.2.0"
description = "Healthcare AI fairness compliance toolkit — audit, monitor, and mitigate bias in clinical AI systems"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.10"
authors = [
    {name = "ParityScope", email = "hello@parityscope.com"},
]
keywords = ["ai fairness", "healthcare", "bias", "compliance", "eu ai act", "audit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "reportlab>=4.0",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
fhir = ["fhir.resources>=7.0"]
onnx = ["onnxruntime>=1.16"]
all = [
    "fhir.resources>=7.0",
    "onnxruntime>=1.16",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.scripts]
parityscope = "parityscope.cli.main:cli"
parityscope-api = "parityscope.api.server:run_server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
